rewrite to publish on zmq instead of rabbit
