+++ README.md
@@ -1,7 +1,7 @@
-ceilometer-publisher-rabbitmq
+ceilometer-publisher-zeromq
 =============================
 
-RabbitMQ Publisher for Ceilometer.
+ZeroMQ Publisher for Ceilometer.
 
 
 Goals
@@ -11,21 +11,19 @@ The traditional Ceilometer publisher, which sends data to MySQL or MongoDB did
 not suit our needs, and our attempt at writing directly to libmarquise
 spoolfiles didn't have the desired performance characteristics.
 
-This publisher writes to a RabbitMQ exchange as simply and quickly as possible,
-leaving collection and subsequent processing to a separate process.
+This publisher writes to a collector via ZeroMQ as simply and quickly as possible,
+leaving subsequent processing to a separate process.
 
 
 Data flow
 ---------
 
 1. Openstack produces a JSON message blob
-2. Enters the Ceilometer pipeline either from a Telemetry (Ceilometer) or
-   Compute (Nova?) node
-3. Is published to the `rabbitmq` sink
-4. JSON message blob is enqueued to RabbitMQ (this is also durable; desirable)
-5. Reader process reads from the RabbitMQ exchange, performs any necessary
-   processing, then writes the output to a spoolfile via libmarquise
-6. Spoolfiles are read and shipped to Vaultaire by a marquised process.
+2. Enters the Ceilometer pipeline via a ceilometer agent
+3. ceilometer-publisher-zeromq sends the JSON to a collector
+4. vaultaire-collector-ceilometer performs any necessary processing,
+   then writes the output to a spoolfile via marquise
+5. Spoolfiles are read and shipped to Vaultaire by a marquised process.
 
 
 Installation + Deployment
@@ -33,12 +31,12 @@ Installation + Deployment
 
 1. Install from source.
     ```
-    git clone git@github.com:anchor/ceilometer-publisher-rabbitmq.git
-    cd ceilometer-publisher-rabbitmq
+    git clone git@github.com:anchor/ceilometer-publisher-zeromq.git
+    cd ceilometer-publisher-zeromq
     python setup.py install
     ```
 
-2. Add *rabbitmq://* to the publishers in your `pipeline.yaml`.
+2. Add *zeromq://* to the publishers in your `pipeline.yaml`.
    (This is by default in `/etc/ceilometer/`)
 
    Example `pipeline.yaml`:
@@ -55,7 +53,7 @@ Installation + Deployment
         - name: meter_sink
           transformers:
           publishers:
-              - rabbitmq://
+              - zeromq://
     ```
 
 3. Add the credentials + options you require to your `ceilometer.conf`
@@ -65,11 +63,8 @@ Installation + Deployment
 
     ```
     [DEFAULT]
-    publisher_exchange = "publisher-exchange"
-    publisher_queue = "publisher-queue"
-    publisher_rabbit_user = guest
-    publisher_rabbit_password = guest
-    publisher_rabbit_host = 127.0.0.1
+    publisher_zeromq_host = 127.0.0.1
+    publisher_zeromq_port = 8282
     ```
 
-4. Restart the central ceilometer agent (`ceilometer-acentral`)
+4. Restart the central and notification ceilometer agents (`ceilometer-acentral + ceilometer-anotification`)
