+++ ceilometer_publisher_zeromq/__init__.py
@@ -1,5 +1,5 @@
 # -*- encoding: utf-8 -*-
 
-from ceilometer_publisher_rabbitmq.queue import *
+from ceilometer_publisher_zeromq.queue import *
 
 __all__ = ['QueuePublisher']
