+++ requirements.txt
@@ -1,1 +1,1 @@
-kombu
+pyzmq
